split mouse_pos into mouse_x and mouse_y
this commit works towards the goal of relegating SFML to the Player.
PlayerInfo links the Player to the Engine and it does not need to use
SFML.

My goal for this project is to eventually move away from SFML, and
create an environment where you need to rewrite only one class.